+++ DofusProtocol/Messages/Messages/game/chat/ChatServerWithObjectMessage.cs
@@ -1,6 +1,6 @@
 
 
-// Generated on 11/16/2015 14:26:01
+// Generated on 12/20/2015 16:36:47
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -24,7 +24,7 @@ namespace Stump.DofusProtocol.Messages
         {
         }
         
-        public ChatServerWithObjectMessage(sbyte channel, string content, int timestamp, string fingerprint, int senderId, string senderName, int senderAccountId, IEnumerable<Types.ObjectItem> objects)
+        public ChatServerWithObjectMessage(sbyte channel, string content, int timestamp, string fingerprint, double senderId, string senderName, int senderAccountId, IEnumerable<Types.ObjectItem> objects)
          : base(channel, content, timestamp, fingerprint, senderId, senderName, senderAccountId)
         {
             this.objects = objects;
