+++ DofusProtocol/Messages/Messages/game/context/fight/GameFightTurnResumeMessage.cs
@@ -1,6 +1,6 @@
 
 
-// Generated on 11/16/2015 14:26:04
+// Generated on 12/20/2015 16:36:49
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -24,7 +24,7 @@ namespace Stump.DofusProtocol.Messages
         {
         }
         
-        public GameFightTurnResumeMessage(int id, int waitTime, int remainingTime)
+        public GameFightTurnResumeMessage(double id, int waitTime, int remainingTime)
          : base(id, waitTime)
         {
             this.remainingTime = remainingTime;
