+++ DofusProtocol/Messages/Messages/game/context/fight/challenge/ChallengeTargetsListMessage.cs
@@ -1,6 +1,6 @@
 
 
-// Generated on 11/16/2015 14:26:04
+// Generated on 12/20/2015 16:36:50
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -18,14 +18,14 @@ namespace Stump.DofusProtocol.Messages
             get { return Id; }
         }
         
-        public IEnumerable<int> targetIds;
+        public IEnumerable<double> targetIds;
         public IEnumerable<short> targetCells;
         
         public ChallengeTargetsListMessage()
         {
         }
         
-        public ChallengeTargetsListMessage(IEnumerable<int> targetIds, IEnumerable<short> targetCells)
+        public ChallengeTargetsListMessage(IEnumerable<double> targetIds, IEnumerable<short> targetCells)
         {
             this.targetIds = targetIds;
             this.targetCells = targetCells;
@@ -38,7 +38,7 @@ namespace Stump.DofusProtocol.Messages
             writer.WriteUShort(0);
             foreach (var entry in targetIds)
             {
-                 writer.WriteInt(entry);
+                 writer.WriteDouble(entry);
                  targetIds_count++;
             }
             var targetIds_after = writer.Position;
@@ -64,10 +64,10 @@ namespace Stump.DofusProtocol.Messages
         public override void Deserialize(IDataReader reader)
         {
             var limit = reader.ReadUShort();
-            var targetIds_ = new int[limit];
+            var targetIds_ = new double[limit];
             for (int i = 0; i < limit; i++)
             {
-                 targetIds_[i] = reader.ReadInt();
+                 targetIds_[i] = reader.ReadDouble();
             }
             targetIds = targetIds_;
             limit = reader.ReadUShort();
