+++ DofusProtocol/Types/Types/game/context/fight/FightResultMutantListEntry.cs
@@ -1,6 +1,6 @@
 
 
-// Generated on 11/16/2015 14:20:21
+// Generated on 12/20/2015 17:30:55
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -23,7 +23,7 @@ namespace Stump.DofusProtocol.Types
         {
         }
         
-        public FightResultMutantListEntry(short outcome, sbyte wave, Types.FightLoot rewards, int id, bool alive, short level)
+        public FightResultMutantListEntry(short outcome, sbyte wave, Types.FightLoot rewards, double id, bool alive, short level)
          : base(outcome, wave, rewards, id, alive)
         {
             this.level = level;
