+++ DofusProtocol/Types/Types/game/context/fight/GameFightFighterTaxCollectorLightInformations.cs
@@ -1,6 +1,6 @@
 
 
-// Generated on 11/16/2015 14:20:21
+// Generated on 12/20/2015 17:30:56
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -24,7 +24,7 @@ namespace Stump.DofusProtocol.Types
         {
         }
         
-        public GameFightFighterTaxCollectorLightInformations(bool sex, bool alive, int id, sbyte wave, short level, sbyte breed, short firstNameId, short lastNameId)
+        public GameFightFighterTaxCollectorLightInformations(bool sex, bool alive, double id, sbyte wave, short level, sbyte breed, short firstNameId, short lastNameId)
          : base(sex, alive, id, wave, level, breed)
         {
             this.firstNameId = firstNameId;
